Simplify things pending further development
I don't use this for anything and at this point there are likely much
better options for handling this sort of stuff. Not sure if I'll put
any more time toward this in the future.

Regardless, if I were to it would not employ Grunt or CoffeeScript or
any of the other crap that was here before. As such I'm removing all
that garbage. Mostly because I need to clear my alerts of all these
stale dev dependencies.